+++ src/Api/Discovery/DiscoveryManager.php
@@ -31,11 +31,11 @@ interface DiscoveryManager
     /**
      * Adds a new binding type.
      *
-     * @param BindingTypeDescriptor $bindingType The binding type to add.
+     * @param BindingTypeDescriptor $typeDescriptor The type to add.
      *
      * @throws DuplicateTypeException If the type is already defined.
      */
-    public function addBindingType(BindingTypeDescriptor $bindingType);
+    public function addBindingType(BindingTypeDescriptor $typeDescriptor);
 
     /**
      * Removes a binding type.
@@ -59,12 +59,9 @@ interface DiscoveryManager
     /**
      * Adds a new binding.
      *
-     * @param        $query
-     * @param        $typeName
-     * @param array  $parameterValues
-     * @param string $language
+     * @param BindingDescriptor $bindingDescriptor The binding to add.
      */
-    public function addBinding($query, $typeName, array $parameterValues = array(), $language = 'glob');
+    public function addBinding(BindingDescriptor $bindingDescriptor);
 
     /**
      * Removes a binding.
@@ -86,6 +83,17 @@ interface DiscoveryManager
      */
     public function enableBinding(Uuid $uuid, $packageName = null);
 
+    /**
+     * Disables a binding.
+     *
+     * @param Uuid            $uuid        The UUID of the binding.
+     * @param string|string[] $packageName The package name to disable the
+     *                                     binding in. Useful if the same
+     *                                     binding exists in multiple packages.
+     *
+     * @throws NoSuchBindingException If the binding could not be found.
+     * @throws CannotDisableBindingException If the binding could not be disabled.
+     */
     public function disableBinding(Uuid $uuid, $packageName = null);
 
     /**